fix: [refactor] ファイル名の変更に伴う依存ファイルの参照を変更
